test: add integration tests for YEK graduation request handling and validation
